(prediction-enhancement): enhance performance

+++ src/Detect/NumClassifier.cpp
@@ -4,15 +4,20 @@
 #include "DataDesc.hpp"
 #include "ExceptionProbe.hpp"
 #include "Hub.hpp"
-#include <caf/event_based_actor.hpp>
 #include <cmath>
 #include <cstdint>
+
+#include "SuppressWarningBegin.hpp"
+
+#include <caf/event_based_actor.hpp>
 #include <fmt/format.h>
 #include <inference_engine.hpp>
 
+#include "SuppressWarningEnd.hpp"
+
 struct NumClassifierSettings final {
-    int32_t inputWidth;
-    int32_t inputHeight;
+    uint32_t inputWidth;
+    uint32_t inputHeight;
     std::string xmlPath;
     std::string binPath;
     std::string deviceName;
@@ -42,8 +47,8 @@ class NumClassifier final : public HubHelper<caf::event_based_actor, NumClassifi
 
         const auto inputData = inputBlob->buffer().as<IE::PrecisionTrait<IE::Precision::FP32>::value_type*>();
 
-        for(size_t h = 0; h < mConfig.inputHeight; h++) {
-            for(size_t w = 0; w < mConfig.inputWidth; w++) {
+        for(uint32_t h = 0; h < mConfig.inputHeight; h++) {
+            for(uint32_t w = 0; w < mConfig.inputWidth; w++) {
                 inputData[h * mConfig.inputWidth + w] = static_cast<float>(dstImg.at<uchar>(h, w)) / 255.0f;
             }
         }
